# docs/swagger.yaml
openapi: 3.0.3
info:
  title: todolist
  version: 0.0.1
  description:  todolist API
servers:
  - url: localhost:3050
paths:
...
  '/user':
    post:
      tags:
        - user
      summary: アカウント作成
      operationId: postUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSimple"
      responses:
        '200':
          description: 作成されたアカウント
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithoutPass"
  /user/{uid}:
    patch:
      tags:
        - user
      summary: アカウント情報編集
      operationId: patchUser
      parameters:
        - name: uid
          in: path
          required: true
          description: ユーザーID
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUser"
      responses:
        '200':
          description: 編集されたアカウント
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithoutPass"
        '404':
          description: 存在しないアカウントです
    delete:
      tags:
        - user
      summary: アカウント削除
      operationId: deleteUser
      parameters:
        - name: uid
          in: path
          required: true
          description: ユーザーID
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: "******"
              required:
                - password
      responses:
        '200':
          description: アカウント削除に成功しました
        '404':
          description: 存在しないアカウントです
...
components:
  schemas:
    UserSimple:
      type: object
      properties:
        userName:
          type: string
          example: testUserName
        password:
          type: string
          example: "********"
      required:
        - userName
        - password
    UserWithoutPass:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        userName:
          type: string
          example: testUserName
      required:
        - userId
        - userName
    UpdateUser:
      type: object
      properties:
        userName:
          type: string
          example: testUserName
        password:
          type: string
          example: "********"
        newPassword:
          type: string
          example: "********"
      required:
        - userName
        - password
        - newPassword
...